Refactoring the structure to a more workable tree
