+++ lib/dashbook.dart
@@ -1,3 +1,3 @@
-export './widget.dart';
-export './story.dart';
-export './decorator.dart';
+export './src/widgets/widget.dart' show Dashbook;
+export './src/story.dart';
+export './src/decorator.dart';
