+++ lib/src/decorator.dart
