+++ lib/src/widgets/property_widgets/bool_property.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../story.dart';
-import 'widgets/property_scaffold.dart';
+import '../../story.dart';
+import './widgets/property_scaffold.dart';
 
 class BoolProperty extends StatefulWidget {
   final Property<bool> property;
