+++ lib/src/widgets/property_widgets/border_radius_property.dart
@@ -1,8 +1,8 @@
-import 'package:dashbook/property_widgets/widgets/property_4_integer_form.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
-import '../story.dart';
-import 'widgets/property_scaffold.dart';
+import './widgets/property_4_integer_form.dart';
+import './widgets/property_scaffold.dart';
+import '../../story.dart';
 
 class BorderRadiusProperty extends StatefulWidget {
   final Property<BorderRadius> property;
