+++ lib/src/widgets/property_widgets/color_property.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
-import '../story.dart';
-import 'widgets/property_scaffold.dart';
-import 'widgets/property_dialog.dart';
+import './widgets/property_scaffold.dart';
+import './widgets/property_dialog.dart';
+import '../../story.dart';
 
 class ColorProperty extends StatefulWidget {
   final Property<Color> property;
@@ -40,7 +40,7 @@ class ColorPropertyState extends State<ColorProperty> {
             pickerAreaHeightPercent: 0.8,
           ),
           actions: [
-            FlatButton(
+            ElevatedButton(
               child: const Text('Got it'),
               onPressed: () {
                 setState(() => currentColor = pickerColor);
