+++ lib/src/widgets/property_widgets/list_property.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../story.dart';
-import 'widgets/property_scaffold.dart';
+import './widgets/property_scaffold.dart';
+import '../../story.dart';
 
 class ListPropertyWidget<T> extends StatefulWidget {
   final ListProperty<T> property;
