+++ lib/src/widgets/property_widgets/number_property.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import '../story.dart';
-import 'widgets/property_scaffold.dart';
+import './widgets/property_scaffold.dart';
+import '../../story.dart';
 
 class NumberProperty extends StatefulWidget {
   final Property<double> property;
